requirements.txt
gradio
groq
Pillow
easyocr
loguru
firebase-admin